loading of history log moved to prompt.py, small edits
